# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Wallet API
  description: Personal finance management REST API for tracking accounts, transactions, budgets, and goals
  version: 1.0.0
  contact:
    name: Wallet
    url: https://github.com/Americo91/wallet
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/transactions/{transactionId}:
    get:
      tags: [Transactions]
      operationId: getTransactionById
      summary: Get transaction by ID
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDto"
        "400":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    put:
      tags: [Transactions]
      operationId: updateTransaction
      summary: Update a transaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionDto"
      responses:
        "204":
          description: Transaction updated
        "400":
          description: Validation error or transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    delete:
      tags: [Transactions]
      operationId: deleteTransaction
      summary: Delete a transaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "204":
          description: Transaction deleted
        "400":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /api/v1/accounts/{accountId}/transactions:
    get:
      tags: [Transactions]
      operationId: listTransactionsByAccount
      summary: List all transactions for an account
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionDto"
        "400":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Create a new transaction for an account
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionDto"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDto"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /api/v1/accounts/{fromAccountId}/transfer/{toAccountId}:
    post:
      tags: [Transactions]
      operationId: transferBetweenAccounts
      summary: Transfer funds between two accounts
      description: Creates paired EXPENSE and INCOME transactions across the source and destination accounts.
      parameters:
        - name: fromAccountId
          in: path
          required: true
          description: Source account UUID
          schema:
            type: string
            format: uuid
        - name: toAccountId
          in: path
          required: true
          description: Destination account UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferDto"
      responses:
        "201":
          description: Transfer completed (returns the two created transactions)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionDto"
        "400":
          description: Validation error or account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"

  # ── Categories ────────────────────────────────────────────────
...
components:
  parameters:
...
    AccountId:
      name: accountId
      in: path
      required: true
      description: Account UUID
      schema:
        type: string
        format: uuid
    TransactionId:
      name: transactionId
      in: path
      required: true
      description: Transaction UUID
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorDto:
      type: object
      properties:
        message:
          type: string
          example: "Institution not found"
      required:
        - message
...
    TransactionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: "c3d4e5f6-a7b8-9012-cdef-123456789012"
        account:
          type: string
          format: uuid
          readOnly: true
          example: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
        type:
          type: string
          enum: [EXPENSE, INCOME]
          example: "EXPENSE"
        amount:
          type: number
          example: 42.50
        date:
          type: string
          format: date
          example: "2025-06-15"
        category:
          type: string
          format: uuid
          example: "d4e5f6a7-b8c9-0123-defa-234567890123"
        description:
          type: [string, "null"]
          example: "Weekly groceries"
        payee:
          type: [string, "null"]
          example: "Supermarket"
        labels:
          type: [array, "null"]
          items:
            type: string
            format: uuid
          uniqueItems: true
          example: ["e5f6a7b8-c9d0-1234-efab-345678901234"]
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-06-15T14:20:00Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-06-15T14:20:00Z"
      required:
        - type
        - amount
        - category
    TransferDto:
      type: object
      properties:
        amount:
          type: number
          example: 500.00
        date:
          type: string
          format: date
          example: "2025-06-15"
        description:
          type: [string, "null"]
          example: "Monthly savings transfer"
        payee:
          type: [string, "null"]
          example: null
        labels:
          type: [array, "null"]
          items:
            type: string
            format: uuid
          uniqueItems: true
          example: null
      required:
        - amount
...